Show only latest SR found base on its date/time
